refactor: MemberRegion 테스트 리팩토링
