+++ second-hand-b/src/test/java/com/codesquad/secondhand/adapter/in/web/MemberRegionSteps.java
@@ -45,8 +45,26 @@ public class MemberRegionSteps {
         Assertions.assertAll(
                 () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
                 () -> assertThat(response.jsonPath().getLong("selectedRegionId")).isEqualTo(3L),
-                () -> assertThat(response.jsonPath().getList("regions")).hasSize(3),
-                () -> assertThat(response.jsonPath().getList("regions.id")).contains(1, 2)
+                () -> assertThat(response.jsonPath().getList("regions")).hasSize(2),
+                () -> assertThat(response.jsonPath().getList("regions.id")).contains(1, 3)
+        );
+    }
+
+    public static ExtractableResponse<Response> 멤버의_지역을_선택한다(Long memberId, Long regionId, String accessToken) {
+        Map<String, Object> body = Map.of("id", regionId);
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header("Authorization", "Bearer " + accessToken)
+                .body(body)
+                .when().put("/api/members/{memberId}/regions", memberId)
+                .then().log().all().extract();
+    }
+
+    public static void 멤버의_지역선택을_검증한다(Long memberId, String accessToken, ExtractableResponse<Response> response) {
+        var memberRegionList = 멤버의_지역목록을_조회한다(memberId, accessToken);
+        Assertions.assertAll(
+                () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
+                () -> assertThat(memberRegionList.jsonPath().getLong("selectedRegionId")).isEqualTo(2)
         );
     }
 }
